JinyoungOh/fastapi/backend/requirements.txt
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
